% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_operator.R
\name{f_operator}
\alias{f_operator}
\title{f operater}
\usage{
f_operator(f)
}
\arguments{
\item{f}{a function}
}
\value{
a function
}
\description{
f operater
}
\examples{
my.log.3 = f_operator(my.log.2)
lapply(runif(2,-10,10), my.log.3)

}
